+++ Automate/Framework/GenericObjectMachine.cs
@@ -1,4 +1,3 @@
-using Microsoft.Xna.Framework;
 using StardewValley;
 using SObject = StardewValley.Object;
 
@@ -37,13 +36,6 @@ namespace Pathoschild.Stardew.Automate.Framework
         protected GenericObjectMachine(TMachine machine, GameLocation location)
             : base(machine, location, BaseMachine.GetTileAreaFor(machine)) { }
 
-        /// <summary>Construct an instance.</summary>
-        /// <param name="machine">The underlying machine.</param>
-        /// <param name="location">The machine's in-game location.</param>
-        /// <param name="tileArea">The tile area covered by the machine.</param>
-        protected GenericObjectMachine(TMachine machine, GameLocation location, in Rectangle tileArea)
-            : base(machine, location, tileArea) { }
-
         /// <summary>Reset the machine so it's ready to accept a new input.</summary>
         /// <param name="item">The output item that was taken.</param>
         protected void GenericReset(Item item)
@@ -55,9 +47,9 @@ namespace Pathoschild.Stardew.Automate.Framework
         /// <summary>Generic logic to pull items from storage based on the given recipes.</summary>
         /// <param name="storage">The available items.</param>
         /// <param name="recipes">The recipes to match.</param>
-        protected bool GenericPullRecipe(IStorage storage, Recipe[] recipes)
+        protected bool GenericPullRecipe(IStorage storage, IRecipe[] recipes)
         {
-            if (storage.TryGetIngredient(recipes, out IConsumable consumable, out Recipe recipe))
+            if (storage.TryGetIngredient(recipes, out IConsumable consumable, out IRecipe recipe))
             {
                 this.Machine.heldObject.Value = recipe.Output(consumable.Take());
                 this.Machine.MinutesUntilReady = recipe.Minutes;