+++ Automate/Framework/Machines/Objects/KegMachine.cs
@@ -12,7 +12,7 @@ namespace Pathoschild.Stardew.Automate.Framework.Machines.Objects
         ** Properties
         *********/
         /// <summary>The recipes to process.</summary>
-        private readonly Recipe[] Recipes =
+        private readonly IRecipe[] Recipes =
         {
             // honey => mead
             new Recipe(